+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Improve relationship with parent and child instances. Refs UIIN-1044.
 * Move items among holdings, even if the list of holdings is loooooooooong. Refs UIIN-1446.
 * Allow the `limit` for the locations query to be specified via stripes config. Refs UIIN-1480.
+* Add visual display when holdings record is suppressed from discovery. Refs UIIN-1378.
+* Add visual display when instance is suppressed from discovery. Refs UIIN-1377 and UIIN-1386.
 * Use the `contributorsNames` index, available in `inventory` since `10.10`. Refs UIIN-1451.
 
 ## [6.0.0](https://github.com/folio-org/ui-inventory/tree/v6.0.0) (2021-03-18)